sys-apps: add usb-modeswitch and usb-modeswitch-data to solve usb dongle issues
As https://packages.debian.org/bullseye/usb-modeswitch-data describes:
Several new USB devices have their proprietary Windows drivers onboard,
especially WAN dongles. When plugged in for the first time, they act
like a flash storage and start installing the driver from there. If the
driver is already installed, the storage device vanishes and a new
device, such as an USB modem, shows up. This is called the "ZeroCD"
feature.

On Debian, this is not needed, since the driver is included as a Linux
kernel module, such as "usbserial". However, the device still shows up
as "usb-storage" by default. usb-modeswitch solves that issue by sending
the command which actually performs the switching of the device from
"usb-storage" to "usbserial".

Add the two packages to solve usb dongle issues.

Signed-off-by: Su Yue <glass@fydeos.io>