Prevent vim ebuild set datadir to /usr/local/share
We intentionally added vim to the dependencies, and vim will be installed
in the root partition. Therefore, use the default `/usr/share` for datadir.
If changed to `/usr/local/share`, running vim will show an error: "Failed to source defaults.vim."

Adding such a USE flag `cros_host` for vim might seem weird,
but it's the minimal change required to achieve the above goal.